\name{KidneyCancerSubtypes-package}
\alias{KidneyCancerSubtypes-package}
\alias{KidneyCancerSubtypes}
\docType{package}
\title{Includes functions to call kidney cancer subtypes including clearcode34, IMDC etc}
\description{Clear code 34 has two subtypes ccA, ccB 
}
\details{
\tabular{ll}{
Package: \tab tt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-22\cr
License: \tab What license is it under?\cr
}

}
\author{Aedin Culhane
Maintainer: <aedin@jimmy.harvard.edu>
}
\references{ references 
}

\keyword{ package }
\seealso{
}
\examples{
}
